move ioports.h

+++ kernel/arch/arm/include/bsp/ioports.h
@@ -1,11 +1,11 @@
 #include <stdint.h>
 
-static inline void iowrite32(uint32_t reg, uint32_t data)
+static inline void iowrite32(uintptr_t reg, uint32_t data)
 {
     *(volatile uint32_t *)reg = data;
 }
 
-static inline uint32_t ioread32(uint32_t reg)
+static inline uint32_t ioread32(uintptr_t reg)
 {
     return *(volatile uint32_t *)reg;
 }
